+++ Server/src/GameWorld/Weapon/Scout.cpp
@@ -5,7 +5,7 @@ Scout::Scout() : Weapon(CF::scout_damage,
                         CF::scout_capacity) {}
 
 void Scout::shoot(std::shared_ptr<Collidable>& player, bool right,
-                  std::map<uint8_t, std::shared_ptr<Infected>>& infecteds) {
+                  std::map<int, std::shared_ptr<Infected>>& infecteds) {
     if (activated) {
         if (right) {
             //atacar a los q tan a la derecha...
